Added one more test and moved the code around
